+++ app/screens/NewTaskScreen/index.jsx
@@ -11,6 +11,7 @@ import { useRouter } from 'expo-router';
 import { getCategories, getTasks, saveTask } from '../../library/storage';
 import { Picker } from '@react-native-picker/picker';
 import { Alert } from 'react-native';
+import { useTask } from '../../library/hooks/taskHooks';
 
 const NewTaskScreen = () => {
   const [taskDescription, setTaskDescription] = useState("");
@@ -20,11 +21,12 @@ const NewTaskScreen = () => {
   const [selectedCategory, setSelectedCategory] = useState(0);
   const [selectedDate, setSelectedDate] = useState(null);
   const [selectedTime, setSelectedTime] = useState(null);
-
+  const { handleSaveTask } = useTask(taskname, taskDescription, selectedCategory, selectedDate, selectedTime, router);
+  
   useEffect(() => {
     const loadCategories = async () => {
       try {
-        const categoriesFromStorage = await getCategories(); // Use your fetch function
+        const categoriesFromStorage = await getCategories(); // Use fetch function
         setCategories(categoriesFromStorage); // Set categories to state
       } catch (error) {
         console.error('Error fetching categories:', error);
@@ -33,47 +35,6 @@ const NewTaskScreen = () => {
 
     loadCategories();
   }, []); // Empty array ensures the effect runs once after the initial render
-
-
-  const handleSaveTask = async () => {
-    // Check if task fields are empty
-    if (!taskname.trim() || !taskDescription.trim() || !selectedCategory) {
-      Alert.alert("Please fill in all fields (Title, Description, and Category).");
-      return; // Prevent saving the task
-    }
-  
-    try {
-      // Retrieve the last used task ID from AsyncStorage
-      const lastTaskId = await AsyncStorage.getItem('lastTaskId');
-      const newTaskId = lastTaskId ? parseInt(lastTaskId) + 1 : 1;
-  
-      // Create new task with a consecutive ID
-      const newTask = {
-        id: newTaskId,
-        title: taskname,
-        description: taskDescription,
-        date: selectedDate.toISOString(),
-        time: selectedTime.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" }),
-        categoryId: selectedCategory, // Ensure the task is linked to a category
-      };
-  
-      // Save task to storage
-      await saveTask(newTask);
-      Alert.alert("Success", "Task saved successfully!");
-  
-      // Update and persist the last used task ID
-      await AsyncStorage.setItem('lastTaskId', newTaskId.toString());
-  
-      // Fetch updated tasks list
-      const tasks = await getTasks();
-  
-      // Redirect to home page
-      router.push('/');
-    } catch (error) {
-      Alert.alert("Error", "There was an issue saving the task. Please try again.");
-    }
-  };
-  
     
   return (
     <Layout>
